Conversations and new rendering mode.
This commit adds the concept of conversations which can be navigated
using C-n and C-p.

It also fundamentally changes the rendering of the program, leaning on
more primitive terminal components so that normal terminal interactions
like selecting text and scrollback are possible.

Additionally there is now message history in the prompt widget that can
be navigated using the up and down keys.